Ability to continue where it left off
I had problems with the Flickr servers and had to write in some logic to allow me to start over without losing any progress.  This could also be useful if someone has to cancel their download for some reason.  The key features are using a central download folder and skipping previously downloaded files.  This can be improved by having a more universal means to determine the file was downloaded.  Also the responsibility of having the values perfected can be removed by adding the options to the preferences.

Other changes were done to facilitate the broader requirements.  Like not creating the file stream until the file successfully downloaded.  And having the option to not create a separate dated download folder for each download attempt (especially useful for Downloading All).  The progress class was modified slightly so that I could see a more granular progress in a large library.

My only suggestions beyond that would be to somehow cache the data requested right before the download begins.